requirements.txt
Flask
transformers
torch
Pillow
opencv-python
numpy
PyJWT
python-dotenv